frontend/requirements.txt
# frontend/requirements.txt
streamlit>=1.24.0
requests>=2.31.0
pillow>=9.5.0,<10.0.0  # Changed to be compatible with streamlit
python-dotenv>=1.0.0